requirements.txt
#torch
#torchvision
pillow
lz4
matplotlib
pygame
opencv-python
